// src/app/shared/utils/enums/estados-cita-informe.ts
export const enum EstadosCitaInformeEnum {
  ACTIVA = 'Activa',
  ELIMINADA = 'Eliminada',
  CUIDADOR = 'Cuidador',
  COMPLETA = 'Completa',
  DESPACHADA = 'Despachada',
  FALLIDA = 'Fallida',
  ENSITIO = 'EnSitio',
  ADMITIDA = 'Admitida',
  ENPELIGRO = 'EnPeligro',
  AGENDADA = 'Agendada',
  ENVIAJE = 'EnViaje',
  INACTIVA = 'Inactiva'
}